+++ source/frontend/widgets/utilities.cpp
