+++ source/frontend/widgets/utilities.hpp
